Add files via upload
A software improvement that prevents the system from locking up due to activity on the serial port input during the boot-up process. It’s not required but recommended to install Raspberry Pi software version 2.7 or higher in addition to the rework to add the pullup on the serial port input.